+++ crates/core/src/kakarot/mod.rs
@@ -1,13 +1,13 @@
 use std::marker::PhantomData;
 
+use reth_primitives::Bytes;
 use starknet::core::types::{BlockId, FunctionCall};
 use starknet::providers::Provider;
 use starknet_crypto::FieldElement;
 
-use crate::client::constants::selectors::{BALANCE_OF, BYTECODE, COMPUTE_STARKNET_ADDRESS, ETH_CALL, GET_EVM_ADDRESS};
-use crate::client::constants::STARKNET_NATIVE_TOKEN;
+use crate::client::constants::selectors::{COMPUTE_STARKNET_ADDRESS, ETH_CALL};
 use crate::client::errors::EthApiError;
-use crate::client::helpers::DataDecodingError;
+use crate::client::helpers::{decode_eth_call_return, vec_felt_to_bytes, DataDecodingError};
 
 pub struct KakarotContract<P> {
     pub address: FieldElement,
@@ -19,42 +19,6 @@ impl<P: Provider + Send + Sync> KakarotContract<P> {
     #[must_use]
     pub fn new(address: FieldElement, proxy_account_class_hash: FieldElement) -> Self {
         Self { address, proxy_account_class_hash, _phantom: PhantomData }
-    }
-
-    pub async fn balance(
-        &self,
-        provider: &P,
-        starknet_address: &FieldElement,
-        block_id: &BlockId,
-    ) -> Result<FieldElement, EthApiError<P::Error>> {
-        // Prepare the calldata for the balanceOf function call
-        let calldata = vec![*starknet_address];
-        let request = FunctionCall {
-            contract_address: FieldElement::from_hex_be(STARKNET_NATIVE_TOKEN).unwrap(),
-            entry_point_selector: BALANCE_OF,
-            calldata,
-        };
-
-        // Make the function call to get the balance
-        let result = provider.call(request, block_id).await?;
-        Ok(*result.first().ok_or_else(|| DataDecodingError::InvalidReturnArrayLength {
-            entrypoint: "balance".into(),
-            expected: 1,
-            actual: 0,
-        })?)
-    }
-
-    pub async fn bytecode(
-        &self,
-        provider: &P,
-        starknet_address: &FieldElement,
-        block_id: &BlockId,
-    ) -> Result<Vec<FieldElement>, EthApiError<P::Error>> {
-        // Prepare the calldata for the bytecode function call
-        let request =
-            FunctionCall { contract_address: *starknet_address, entry_point_selector: BYTECODE, calldata: vec![] };
-
-        Ok(provider.call(request, block_id).await?)
     }
 
     pub async fn compute_starknet_address(
@@ -87,37 +51,23 @@ impl<P: Provider + Send + Sync> KakarotContract<P> {
         eth_address: &FieldElement,
         eth_calldata: &mut Vec<FieldElement>,
         block_id: &BlockId,
-    ) -> Result<Vec<FieldElement>, EthApiError<P::Error>> {
+    ) -> Result<Bytes, EthApiError<P::Error>> {
         let mut calldata =
             vec![*eth_address, FieldElement::MAX, FieldElement::ZERO, FieldElement::ZERO, eth_calldata.len().into()];
 
         calldata.append(eth_calldata);
 
         let request = FunctionCall { contract_address: self.address, entry_point_selector: ETH_CALL, calldata };
-        Ok(provider.call(request, block_id).await?)
-    }
-
-    pub async fn get_evm_address(
-        &self,
-        provider: &P,
-        starknet_address: &FieldElement,
-        block_id: &BlockId,
-    ) -> Result<FieldElement, EthApiError<P::Error>> {
-        let request = FunctionCall {
-            contract_address: *starknet_address,
-            entry_point_selector: GET_EVM_ADDRESS,
-            calldata: vec![],
-        };
-
         let result = provider.call(request, block_id).await?;
-        match result.first() {
-            Some(x) if result.len() == 1 => Ok(*x),
-            _ => Err(DataDecodingError::InvalidReturnArrayLength {
-                entrypoint: "get_evm_address".into(),
-                expected: 1,
-                actual: 0,
-            }
-            .into()),
-        }
+
+        // Parse and decode Kakarot's call return data (temporary solution and not scalable - will
+        // fail is Kakarot API changes)
+        // Declare Vec of Result
+        // TODO: Change to decode based on ABI or use starknet-rs future feature to decode return
+        // params
+        let return_data = decode_eth_call_return(&result)?;
+
+        let result = vec_felt_to_bytes(return_data);
+        Ok(result)
     }
 }
